Move icon command into own file
